correciones 2(? sigue sin funcionar

+++ src/main/java/goingto/com/controller/CountriesController.java
@@ -30,7 +30,7 @@ public class CountriesController {
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<Country>getById(@PathVariable Long id)
+	public ResponseEntity<Country>getById(@PathVariable Integer id)
 	{
 		Country country = countryService.getCountry(id);
 		if(country ==null)
