+++ src/main/java/goingto/com/repository/ICityRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import goingto.com.model.City;
 import goingto.com.model.Country;
 
-public interface ICityRepository extends JpaRepository<City, Long> {
+public interface ICityRepository extends JpaRepository<City, Integer> {
  List<City> findByCountryId(Country country);
 }
