+++ src/main/java/goingto/com/repository/ICountryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import goingto.com.model.Country;
 
-public interface ICountryRepository extends JpaRepository<Country, Long> {
+public interface ICountryRepository extends JpaRepository<Country, Integer> {
 
 }
