+++ src/main/java/goingto/com/repository/ILocatableRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sun.jdi.Locatable;
 
-public interface ILocatableRepository extends JpaRepository<Locatable, Long> {
+public interface ILocatableRepository extends JpaRepository<Locatable, Integer> {
 
 }
